refactor(habitat): drain public surface guard tests
